Rename layer.contracts.assets imports to asset
