+++ layer/contracts/tracker.py
@@ -6,7 +6,7 @@ from typing import Any, Dict, Optional, Union
 
 from yarl import URL
 
-from layer.contracts.assets import AssetType
+from layer.contracts.asset import AssetType
 
 
 class ResourceTransferState:
