+++ layer/projects/execution_planner_old.py
@@ -13,7 +13,7 @@ from layerapi.api.entity.operations_pb2 import (
 )
 from layerapi.api.ids_pb2 import ModelVersionId
 
-from layer.contracts.assets import AssetPath, AssetType
+from layer.contracts.asset import AssetPath, AssetType
 from layer.contracts.definitions import FunctionDefinition
 from layer.exceptions.exceptions import (
     LayerClientException,
