+++ test/unit/decorators/test_model_decorator.py
@@ -3,7 +3,7 @@ from typing import Any, Callable
 import pytest
 
 from layer import Dataset, Model
-from layer.contracts.assets import AssetType
+from layer.contracts.asset import AssetType
 from layer.decorators.model_decorator import model
 from layer.decorators.pip_requirements_decorator import pip_requirements
 from layer.exceptions.exceptions import ProjectInitializationException
